Adiciona nova funcionalidade de Respostas, documenta a api e arruma erros de teste de endpoint

+++ src/main/java/br/com/api/forum_hub/controllers/CourseController.java
@@ -1,17 +1,14 @@
 package br.com.api.forum_hub.controllers;
 
-import br.com.api.forum_hub.dtos.CreateCourseDTO;
-import br.com.api.forum_hub.dtos.RegisterTopicDTO;
-import br.com.api.forum_hub.dtos.ResponseTopicDTO;
-import br.com.api.forum_hub.dtos.UpdateTopicDTO;
+import br.com.api.forum_hub.dtos.*;
 import br.com.api.forum_hub.models.Course;
-import br.com.api.forum_hub.models.Topic;
 import br.com.api.forum_hub.repositories.CourseRepository;
-import br.com.api.forum_hub.repositories.TopicRepository;
-import br.com.api.forum_hub.services.DeleteTopicUseCase;
-import br.com.api.forum_hub.services.FindTopicUseCase;
-import br.com.api.forum_hub.services.RegisterTopicUseCase;
-import br.com.api.forum_hub.services.UpdateTopicUseCase;
+import br.com.api.forum_hub.services.ValidationException;
+import io.swagger.v3.oas.annotations.enums.SecuritySchemeType;
+import io.swagger.v3.oas.annotations.security.SecurityRequirement;
+import io.swagger.v3.oas.annotations.security.SecurityRequirements;
+import io.swagger.v3.oas.annotations.security.SecurityScheme;
+import jakarta.persistence.EntityNotFoundException;
 import jakarta.transaction.Transactional;
 import jakarta.validation.Valid;
 import org.springframework.http.ResponseEntity;
@@ -23,9 +20,10 @@ import java.util.List;
 
 @RequestMapping("/courses")
 @RestController
+@SecurityRequirement(name = "courses")
 public class CourseController {
 
-    private CourseRepository repository;
+    private final CourseRepository repository;
 
     public CourseController(CourseRepository repository) {
         this.repository = repository;
@@ -33,26 +31,30 @@ public class CourseController {
 
     @PostMapping
     @Transactional
-    public ResponseEntity<Course> create(@RequestBody @Valid CreateCourseDTO data, UriComponentsBuilder uriBuilder){
+    public ResponseEntity<CourseResponseDTO> create(@RequestBody @Valid CourseRequestDTO data, UriComponentsBuilder uriBuilder){
+        if (repository.existsByName(data.name())) throw new ValidationException("Já existe um curso com esse nome");
+
         Course course = new Course(data);
-        Course savedCourse = repository.save(course);
+        CourseResponseDTO courseDto = new CourseResponseDTO(repository.save(course));
 
         URI uri = uriBuilder.path("/courses/{id}")
-                .buildAndExpand(savedCourse.getId())
+                .buildAndExpand(courseDto.id())
                 .toUri();
 
-        return ResponseEntity.created(uri).body(savedCourse);
+        return ResponseEntity.created(uri).body(courseDto);
     }
 
     @GetMapping
-    public ResponseEntity<List<Course>> listAll(){
-        List<Course> allTopics = repository.findAll();
-        return ResponseEntity.ok(allTopics);
+    public ResponseEntity<List<CourseResponseDTO>> listAll(){
+        List<Course> allCourses = repository.findAll();
+        List<CourseResponseDTO> allCoursesDto = allCourses.stream().map(CourseResponseDTO::new).toList();
+        return ResponseEntity.ok(allCoursesDto);
     }
 
 
     @DeleteMapping("/{id}")
     public ResponseEntity<Void> delete(@PathVariable Long id){
+        if(!repository.existsById(id)) throw new EntityNotFoundException("Não existe nenhum curso com esse id");
         repository.deleteById(id);
         return ResponseEntity.noContent().build();
     }
